Created Disaster Director Prefab and Disaster Type Script
Created disaster director script and put it on a gameobject to make a prefab
Created a folder for system prefabs
Created DisasterType script that determines spawning properties of a disaster prefab which are used by the disaster director
DisasterDirector has animation curves to easily scale the difficulty of waves. It despawns disasters when entering a rest period. Uses a credits system to determine what it can or can't spawn. It spawns disasters based on the position of the player